zshrc, zshenv, zloginをシェルで作るようにした

+++ install.sh
@@ -1,29 +1,35 @@
 #!/bin/sh
 
-cat <<'EOF'
+if [ ! -e $HOME/.zsh.d ]; then
+  ln -sF $HOME/.dotfiles/myzsh/zsh.d  $HOME/.zsh.d
+fi
 
-,--.,--.        ,--.    ,--.                            
-|  |`--',--,--, |  |,-. `--',--,--,  ,---.              
-|  |,--.|      \|     / ,--.|      \| .-. |             
-|  ||  ||  ||  ||  \  \ |  ||  ||  |' '-' '.--..--..--. 
-`--'`--'`--''--'`--'`--'`--'`--''--'.`-  / '--''--''--' 
-                                    `---'               
+if [ ! -e $HOME/.zlogin ]; then
+cat <<'EOF' > $HOME/.zlogin
+echo "Hello, Z World !"
 
 EOF
+fi
 
-ln -sF $HOME/.dotfiles/myzsh/zlogin $HOME/.zlogin
-ln -sF $HOME/.dotfiles/myzsh/zshenv $HOME/.zshenv
-ln -sF $HOME/.dotfiles/myzsh/zshrc  $HOME/.zshrc
-ln -sF $HOME/.dotfiles/myzsh/zsh.d  $HOME/.zsh.d
+cat <<'EOF' > $HOME/.zshenv
+source ~/.zsh.d/zshenv
+EOF
+
+cat <<'EOF' > $HOME/.zshrc
+source ~/.zsh.d/zshrc
+source ~/.zsh.d/aliases.zsh
+source ~/.zsh.d/theme.zsh
+source ~/.zsh.d/config/packages.zsh
+EOF
+
 ls -hal ~/.zsh{.d,env,rc} ~/.zlogin
 
 cat <<'EOF'
 
-   ,--.                       
- ,-|  | ,---. ,--,--,  ,---.  
+,--.                       
+,-|  | ,---. ,--,--,  ,---.  
 ' .-. || .-. ||      \| .-. : 
 \ `-' |' '-' '|  ||  |\   --. 
- `---'  `---' `--''--' `----' 
+`---'  `---' `--''--' `----' 
 
 EOF
-
